WIP: allow import to take destination path.
When using --substitute with a stacker file, I came across a stacker
the desire to have a destination path on the import.

I had a stacker file that was using --substitute to get a directory
'import'ed.  But then we wanted to use the path to that directory
inside the 'run' section.

Imagine that the default 'SCRIPTS_DIR' below is named
'/path/to/scripts', and is expected to have a script named 'do-magic'
in it.

  import:
   - ${{SCRIPTS_DIR}}
  run: |
    PATH=/stacker/scripts:$PATH
    echo "running your script"
    do-magic

That will all work well.

However, if we want to change the name of SCRIPTS_DIR to
'common-scripts' then we'd have to change the yaml in two otherwise
unrelated places (import and /run).  It is also unclear from reading
this yaml file where /scripts comes from at all.  It seems annoying at
least that there is an expectation on the content of the 'SCRIPTS_DIR'
variable (it must end in /scripts).

What I'm proposing for a fix here is instead:

  import:
    - ${{SCRIPTS_DIR}} -> /scripts

Then the /scripts string is present in the two places that it is
used and updates to the value of SCRIPTS_DIR do not force changes
to the 'run:' section.

TODO: I need to fix the paths that will be used.
On the right side of -> the path should not depend on whether or
not there existed a directory by that name.  Or maybe it shoudl
but I think there is still work to do in 'Import' function.

Fixes https://github.com/anuvu/stacker/issues/41